+++ src/Payline/Payment.php
@@ -29,4 +29,6 @@ class Payment
     public $softDescriptor;
 
     public $cardBrand;
+
+    public $registrationToken;
 }