+++ src/Payline/SubMerchant.php
@@ -1,0 +1,38 @@
+<?php
+/*
+ * This file is part of the Payline package.
+ *
+ * (c) Monext <http://www.monext.net>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace Payline;
+
+class SubMerchant
+{
+
+    public $subMerchantId;
+
+    public $subMerchantName;
+
+    public $subMerchantMCC;
+
+    public $subMerchantSIRET;
+
+    public $subMerchantTaxCode;
+
+    public $subMerchantStreet;
+
+    public $subMerchantCity;
+
+    public $subMerchantZipCode;
+
+    public $subMerchantCountry;
+
+    public $subMerchantState;
+
+    public $subMerchantEmailAddress;
+
+    public $subMerchantPhoneNumber;
+}